Move borrowed files to external folder
Previously, code for mbedtls and code from skiboot were being mixed with code that was written specifically for this project. In order to give proper credit to code that we did not write, we have put those files into the 'external' subdirectory. Then we edited the paths and build tools so that the package functions the same.
