+++ CMakeLists.txt
@@ -9,21 +9,31 @@ list( TRANSFORM DEPEN PREPEND ${DEPDIR} )
 set( SRC secvarctl.c generic.c )
 
 #sources/dependencies for edk2 backend
-set( SKIBOOTDEPEN list.h config.h container_of.h check_type.h secvar.h opal-api.h endian.h short_types.h edk2.h edk2-svc.h )
-set( SKIBOOTDEPDIR backends/edk2-compat/include/ )
+set( EDK2DEPEN edk2-svc.h )
+set( EDK2DEPDIR backends/edk2-compat/include/ )
+list( TRANSFORM EDK2DEPEN PREPEND ${EDK2DEPDIR} )
+set ( EDK2SRC edk2-svc-read.c edk2-svc-write.c edk2-svc-validate.c edk2-svc-verify.c edk2-svc-generate.c )
+set ( EDK2SRCDIR backends/edk2-compat/ )
+list( TRANSFORM EDK2SRC PREPEND ${EDK2SRCDIR} )
+list( APPEND DEPEN ${EDK2DEPEN} )
+list( APPEND SRC ${EDK2SRC} )
+
+#sources/dependencies for borrowed skiboot code
+set( SKIBOOTDEPEN list.h config.h container_of.h check_type.h secvar.h opal-api.h endian.h short_types.h edk2.h edk2-compat-process.h )
+set( SKIBOOTDEPDIR external/skiboot/include/ )
 list( TRANSFORM SKIBOOTDEPEN PREPEND ${SKIBOOTDEPDIR} )
-set ( SKIBOOTSRC secvar_util.c edk2-svc-read.c edk2-svc-write.c edk2-svc-validate.c edk2-compat.c edk2-compat-process.c   edk2-svc-verify.c edk2-svc-generate.c )
-set ( SKIBOOTSRCDIR backends/edk2-compat/ )
+set ( SKIBOOTSRC secvar_util.c edk2-compat.c edk2-compat-process.c )
+set ( SKIBOOTSRCDIR external/skiboot/ )
 list( TRANSFORM SKIBOOTSRC PREPEND ${SKIBOOTSRCDIR} )
-list( APPEND DEPEN ${SKIBOOTDEP} )
+list( APPEND DEPEN ${SKIBOOTDEPEN} )
 list( APPEND SRC ${SKIBOOTSRC} )
 
 #sources/dependencies for extra mbedtls functions
 set( EXTRAMBEDTLSDEP generate-pkcs7.h pkcs7.h )
-set( EXTRAMBEDTLSDEPDIR extraMbedtls/include/ )
+set( EXTRAMBEDTLSDEPDIR external/extraMbedtls/include/ )
 list( TRANSFORM EXTRAMBEDTLSDEP PREPEND ${EXTRAMBEDTLSDEPDIR} )
 set ( EXTRAMBEDTLSSRC generate-pkcs7.c pkcs7.c )
-set ( EXTRAMBEDTLSSRCDIR  extraMbedtls/ )
+set ( EXTRAMBEDTLSSRCDIR  external/extraMbedtls/ )
 list( TRANSFORM EXTRAMBEDTLSSRC PREPEND ${EXTRAMBEDTLSSRCDIR} )
 list( APPEND DEPEN ${EXTRAMBEDTLSDEP} )
 list( APPEND SRC ${EXTRAMBEDTLSSRC} )
