+++ Makefile
@@ -6,25 +6,36 @@ LFLAGS = -lmbedtls -lmbedx509 -lmbedcrypto -Wall -Werror
 _DEPEN = secvarctl.h prlog.h err.h generic.h 
 DEPDIR = include
 DEPEN = $(patsubst %,$(DEPDIR)/%, $(_DEPEN))
-_SKIBOOT_DEPEN =list.h config.h container_of.h check_type.h secvar.h opal-api.h endian.h short_types.h edk2.h edk2-svc.h 
-SKIBOOTDEPDIR = backends/edk2-compat/include
-SKIBOOTDEPEN = $(patsubst %,$(SKIBOOTDEPDIR)/%, $(_SKIBOOT_DEPEN))
-DEPEN += $(SKIBOOTDEPEN)
+
+_EDK2_DEPEN = edk2-svc.h 
+EDK2DEPDIR = backends/edk2-compat/include
+EDK2_DEPEN = $(patsubst %,$(EDK2DEPDIR)/%, $(_EDK2_DEPEN))
+DEPEN += $(EDK2_DEPEN)
+
+_SKIBOOT_DEPEN =list.h config.h container_of.h check_type.h secvar.h opal-api.h endian.h short_types.h edk2.h edk2-compat-process.h
+SKIBOOTDEPDIR = external/skiboot/include
+SKIBOOT_DEPEN = $(patsubst %,$(SKIBOOTDEPDIR)/%, $(_SKIBOOT_DEPEN))
+DEPEN += $(SKIBOOT_DEPEN)
 
 _EXTRAMBEDTLS_DEPEN = pkcs7.h generate-pkcs7.h 
-EXTRAMBEDTLSDEPDIR = extraMbedtls/include
+EXTRAMBEDTLSDEPDIR = external/extraMbedtls/include
 EXTRAMBEDTLSDEPEN = $(patsubst %,$(EXTRAMBEDTLSDEPDIR)/%, $(_EXTRAMBEDTLS_DEPEN))
 DEPEN += $(EXTRAMBEDTLSDEPEN)
 
-SKIBOOTOBJDIR = backends/edk2-compat
-_SKIBOOT_OBJ = secvar_util.o edk2-svc-read.o edk2-svc-write.o edk2-svc-validate.o edk2-compat.o edk2-compat-process.o   edk2-svc-verify.o edk2-svc-generate.o
+EDK2OBJDIR = backends/edk2-compat
+_EDK2_OBJ =  edk2-svc-read.o edk2-svc-write.o edk2-svc-validate.o edk2-svc-verify.o edk2-svc-generate.o
+EDK2_OBJ = $(patsubst %,$(EDK2OBJDIR)/%, $(_EDK2_OBJ))
+
+SKIBOOTOBJDIR = external/skiboot/
+_SKIBOOT_OBJ = secvar_util.o edk2-compat.o edk2-compat-process.o
 SKIBOOT_OBJ = $(patsubst %,$(SKIBOOTOBJDIR)/%, $(_SKIBOOT_OBJ))
 
-EXTRAMBEDTLSDIR = extraMbedtls
+EXTRAMBEDTLSDIR = external/extraMbedtls
 _EXTRAMBEDTLS = generate-pkcs7.o pkcs7.o 
 EXTRAMBEDTLS = $(patsubst %,$(EXTRAMBEDTLSDIR)/%, $(_EXTRAMBEDTLS))
+
 OBJ =secvarctl.o  generic.o 
-OBJ +=$(SKIBOOT_OBJ) $(EXTRAMBEDTLS)
+OBJ +=$(SKIBOOT_OBJ) $(EXTRAMBEDTLS) $(EDK2_OBJ)
 
 OBJCOV = $(patsubst %.o, %.cov.o,$(OBJ))
 
@@ -42,10 +53,6 @@ endif
 NO_CRYPTO = 0 
 ifeq ($(NO_CRYPTO),1)
 	_CFLAGS+=-DNO_CRYPTO
-else 
-	_CRYPTO_OBJ +=  edk2-svc-generate.o
-	CRYPTO_OBJ = $(patsubst %,$(SKIBOOTOBJDIR)/%, $(_CRYPTO_OBJ))
-	OBJ += $(CRYPTO_OBJ)
 endif
 
 
@@ -59,7 +66,7 @@ secvarctl: $(OBJ)
 
 clean:
 	rm -f $(OBJ) secvarctl 
-	rm -f ./*/*.cov.* secvarctl-cov ./*.cov.* ./backends/*/*.cov.* ./html*
+	rm -f ./*/*.cov.* secvarctl-cov ./*.cov.* ./backends/*/*.cov.* ./external/*/*.cov.* ./html*
 
 
 %.cov.o: %.c $(DEPEN)
