+++ backends/edk2-compat/include/edk2-svc.h
@@ -2,12 +2,12 @@
 #define EDK2_SVC_SKIBOOT_H
 #include <stdint.h> //for uint_16 stuff like that
 #include <mbedtls/x509_crt.h> // for printCertInfo
-#include "secvar.h" //for secvar struct
+#include "../../../external/skiboot/include/secvar.h" //for secvar struct
 #include "../../../include/err.h"
 #include "../../../include/prlog.h"
 #include "../../../include/generic.h"
-#include "../../../extraMbedtls/include/generate-pkcs7.h"
-#include "edk2.h" //include last or else problems from pragma pack(1)
+#include "../../../external/extraMbedtls/include/generate-pkcs7.h"
+#include "../../../external/skiboot/include/edk2.h" //include last or else problems from pragma pack(1)
 
 
 #define CERT_BUFFER_SIZE        2048
