+++ external/skiboot/edk2-compat.c
@@ -20,7 +20,7 @@
 
 
 //added by Nick Child, not official
-#include "../../extraMbedtls/include/pkcs7.h"
+#include "../extraMbedtls/include/pkcs7.h"
 #include "include/edk2.h"
 #include "../../include/prlog.h"
 #include "include/secvar.h"
