+++ external/skiboot/include/edk2.h
