+++ external/skiboot/include/endian.h
