+++ external/skiboot/include/opal-api.h
